Used OpenCv to complete Otsu thresholding method with the guassian blurring inside - depending how we continue with this can always remove this and do separetely but it's only one line so will see
In the main file I've just cleaned up linking to mine - need to standardise the variable names

Also on the advice of what I've seen discuseed online I've added a dependecies.txt file - if you use any new modules that's not already been mentioned inside of it please add. I don't believe the links are necessary in the grand scheme of things but for each other is nice we don't have to jump around google looking for these things.